Logic.Proof

Proof Trees

data Proof j

The type of proof trees parameterized by the type of judgments j.

Constructors

Node j [Proof j]

pattern Proof :: j -> [Proof j] -> Proof j

A constructor synonym for Node.

pattern Leaf :: j -> Proof j

A constructor synonym for Node with no premises.

Proof Generation

class Explain j where

Methods

premises :: j -> [[j]]

Returns a list of all possible premises of a given judgment.

premises j = [] if j is decidably false.A

premises j = [[]] if j is an axiom.

premises j = [p1,p2,...,pn] : tail (premises j) if j is provable from the conclusions of p1, p2, ..., pn.

proofs :: Explain j => j -> [Proof j]

Returns a list of all possible proofs of a given judgment.

prove :: Explain j => j -> Maybe (Proof j)

Returns the first proof of a given judgment.

prove j = listToMaybe (proofs j)

prove' :: Explain j => j -> Proof j

Returns a proof of a given judgment or fails if the judgment is not provable.

prove' j = fromJust (prove j)

suppose :: Explain j => j -> Proof j

Gives the first proof of a given judgment and doesn't fail if the judgement is not provable.

Proof Tree Manipulation

conclusion :: Proof j -> j

Extract the conclusion/root from a proof tree.

children :: Proof j -> [Proof j]

Extract the premises/children from a proof tree.

hidePast :: Int -> Proof j -> Proof j

Get the proof tree with the premises hidden after the \(n\)th level.

hide :: Eq j => j -> Proof j -> Maybe (Proof j)

Hide the subtrees of a proof tree that have the given judgment as their conclusion.

countNodes :: Proof j -> Int

Count the number of judgemnets in a proof tree.

getRow :: Int -> Proof j -> [j]

Get the \(n\)th row of a proof tree.

toList :: Eq j => Proof j -> [j]

Convert the proof tree to a list of judgements.

toList' :: Proof a -> [a]

Convert the proof tree to a list of judgements.